+++ stream-chat-android-docs/src/main/java/io/getstream/chat/docs/java/ui/messages/MessageListViewSnippets.java
@@ -21,6 +21,7 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.concurrent.TimeUnit;
 
+import io.getstream.chat.android.common.state.messagelist.MessagePosition;
 import io.getstream.chat.android.ui.TransformStyle;
 import io.getstream.chat.android.ui.message.list.MessageListView;
 import io.getstream.chat.android.ui.message.list.adapter.BaseMessageItemViewHolder;
@@ -236,6 +237,6 @@ public class MessageListViewSnippets extends Fragment {
     }
 
     public void avatarPredicate() {
-        messageListView.setShowAvatarPredicate((messageItem) -> messageItem.getPositions().contains(MessageListItem.Position.BOTTOM) && messageItem.isTheirs());
+        messageListView.setShowAvatarPredicate((messageItem) -> messageItem.getPositions().contains(MessagePosition.BOTTOM) && messageItem.isTheirs());
     }
 }